TKSS-451: Decouple SM2 operations from ECOperations
