+++ kona-crypto/src/main/java/com/tencent/kona/crypto/provider/SM2Signature.java
@@ -55,9 +55,9 @@ import static com.tencent.kona.crypto.spec.SM2ParameterSpec.CURVE;
 import static com.tencent.kona.crypto.spec.SM2ParameterSpec.GENERATOR;
 import static com.tencent.kona.crypto.spec.SM2ParameterSpec.ORDER;
 import static com.tencent.kona.crypto.util.Constants.defaultId;
-import static com.tencent.kona.sun.security.ec.ECOperations.INFINITY;
-import static com.tencent.kona.sun.security.ec.ECOperations.SM2OPS;
-import static com.tencent.kona.sun.security.ec.ECOperations.toECPoint;
+import static com.tencent.kona.crypto.util.Constants.INFINITY;
+import static com.tencent.kona.sun.security.ec.SM2Operations.SM2OPS;
+import static com.tencent.kona.sun.security.ec.SM2Operations.toECPoint;
 import static java.math.BigInteger.ONE;
 import static java.math.BigInteger.ZERO;
